+++ CHANGELOG.md
@@ -1,3 +1,8 @@
+0.3.7 - 18 Jul 2026
+===================
+- feat: Support client-side and bidirectional streaming (send messages with `Enter`, finish with `Ctrl+d`)
+- feat: Update edtui: Supports dot-repeat and more vim-like keybindings
+
 0.3.5 - 14 Jan 2026
 ===================
 - feat: Add mouse scrolling support